Commit before major change on logic | Changing buttons to AJAX requests
In order to create a better UX, the buttons to send message or new friendship should update their status when clicked dinamically, instead of on page render, so it is necessary to change the buttons from a link_to to a button with a data-action to start an AJAX request, and upon response re-render the specific section with the buttons. Will be necessary to add the 'respond_to' with the JSON format.